css improved

+++ public/js/commentGuest.js
@@ -3,18 +3,19 @@ const deleteComment = document.querySelectorAll('.delete-comment');
 const editComment = document.querySelectorAll('.edit-comment');
 const commentsContent = document.querySelector('.comments-content');
 const commentformWrapper = document.querySelector('.comment-form-wrapper');
-const submitComment = document.querySelector('.comment-form button');
+const submitComment = document.querySelector('#submit-comment');
+const cancelComment = document.querySelector('#cancel-comment');
 const commentContent = document.querySelector('.comment-form textarea');
 const guestInput = document.querySelector('#guest-form input');
 const guestSignUp = document.querySelector('#guest-signup');
 const guestSignout = document.querySelector('#guset-signout');
 
-//Get the data from the user input and fetch the backend api for creating a new comment.
+//Get the data from the user input and fetch the backend api for creating a new comment, or editing.
 const addCommentHandler = async (e) => {
   e.preventDefault();
   const content = commentContent.value.trim();
   const eventId = submitComment.dataset.eventid;
-  const id = document.querySelector('#comment-id').value
+  const id = document.querySelector('#comment-id').value;
 
   if (content) {
     try {
@@ -25,7 +26,17 @@ const addCommentHandler = async (e) => {
       });
 
       if (response.ok) {
-        document.location.reload();
+        if (!id) {
+          document.location.reload();
+        } else {
+          editComment.forEach((el) => {
+            if (el.dataset.commentid === id) {
+              el.previousElementSibling.previousElementSibling.innerHTML = `${content} <span class="text-secondary">(edited)</span>`;
+              commentformWrapper.classList.add('hide');
+              console.log(el.previousElementSibling.previousSibling);
+            }
+          });
+        }
       } else {
         alert('Failed to comment.');
       }
@@ -38,33 +49,38 @@ const addCommentHandler = async (e) => {
 };
 
 const deleteCommentHandler = async (el) => {
-  const id = el.dataset.commentid;
+  const confirm = window.confirm('Are you sure?');
 
-  try {
-    const response = await fetch('/api/comments', {
-      method: 'DELETE',
-      body: JSON.stringify({ id }),
-      headers: { 'Content-Type': 'application/json' },
-    });
+  if (confirm) {
+    const id = el.dataset.commentid;
 
-    if (response.ok) {
-      el.parentElement.remove();
-    } else {
+    try {
+      const response = await fetch('/api/comments', {
+        method: 'DELETE',
+        body: JSON.stringify({ id }),
+        headers: { 'Content-Type': 'application/json' },
+      });
+
+      if (response.ok) {
+        el.parentElement.remove();
+      } else {
+        alert('Failed to Delete.');
+      }
+    } catch (e) {
       alert('Failed to Delete.');
     }
-  } catch (e) {
-    alert('Failed to Delete.');
   }
 };
 
 const editCommentHandler = (el) => {
-
-  // commentsContent.classList.add('hide');
   commentformWrapper.classList.remove('hide');
+  commentContent.value = el.dataset.commentcontent;
+  document.querySelector('#comment-id').value = el.dataset.commentid;
+};
 
-  commentContent.value = el.dataset.commentcontent
-  document.querySelector('#comment-id').value = el.dataset.commentid
-}
+const cancelCommentHandler = () => {
+  commentformWrapper.classList.add('hide');
+};
 
 //Get the data from the user input and fetch the backend api for creating a new guest.
 const addGuestHandler = async (e) => {
@@ -112,9 +128,10 @@ const deleteGuestHandler = async (e) => {
   }
 };
 
-// Toggle between comments form and comments content elements 
+// Toggle between comments form and comments content elements
 addComment.addEventListener('click', () => {
-  // commentsContent.classList.add('hide');
+  document.querySelector('#comment-id').value = '';
+  commentContent.value = '';
   commentformWrapper.classList.remove('hide');
 });
 
@@ -126,11 +143,13 @@ deleteComment.forEach((el) => {
 
 editComment.forEach((el) => {
   el.addEventListener('click', () => {
-    editCommentHandler(el)
+    editCommentHandler(el);
   });
 });
 
 submitComment?.addEventListener('click', addCommentHandler);
+
+cancelComment?.addEventListener('click', cancelCommentHandler);
 
 guestSignUp?.addEventListener('click', addGuestHandler);
